1. 권한그룹 삭제 완료

+++ src/main/java/egovframework/com/devjitsu/model/menu/TblMenuAuthrtGroup.java
@@ -53,7 +53,7 @@ public class TblMenuAuthrtGroup {
 
     @Column(name = "CREATR_SN", columnDefinition = "INT(10)", updatable=false, nullable = false)
     @Comment("생성자일련번호")
-    private int creatrSn;
+    private long creatrSn;
 
     @Column(name = "FRST_CRT_DT", columnDefinition = "DATETIME DEFAULT CURRENT_TIMESTAMP")
     @Comment("최초생성일시")
